requirements.txt
pandas
streamlit
requests
pyyaml